# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ManiocAgri",
    "description": "Backend for ManiocAgri Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/orders/": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Create Order",
        "description": "Create new order.",
        "operationId": "create_order_api_v1_orders__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Read Orders",
        "description": "Retrieve orders.\n- Admins/Managers see all.\n- Drivers see assigned orders.\n- Clients see their own.",
        "operationId": "read_orders_api_v1_orders__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderRead"
                  },
                  "title": "Response Read Orders Api V1 Orders  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/track/{number}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Track Order",
        "description": "Track an order by its number.",
        "operationId": "track_order_api_v1_orders_track__number__get",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/pending": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Read Pending Orders",
        "description": "Get all pending orders that need assignment. Admin/Gestionnaire only.",
        "operationId": "read_pending_orders_api_v1_orders_pending_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderRead"
                  },
                  "title": "Response Read Pending Orders Api V1 Orders Pending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}/status": {
      "patch": {
        "tags": [
          "orders"
        ],
        "summary": "Update Order Status",
        "description": "Update order status. Only for staff.",
        "operationId": "update_order_status_api_v1_orders__id__status_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}/assign": {
      "patch": {
        "tags": [
          "orders"
        ],
        "summary": "Assign Order To Livreur",
        "description": "Assign an order to a delivery person.\nAdmin et gestionnaire only.",
        "operationId": "assign_order_to_livreur_api_v1_orders__id__assign_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "livreur_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Livreur Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/livreurs": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get Available Livreurs",
        "description": "Get list of available delivery persons for assignment.\nAdmin et gestionnaire only.",
        "operationId": "get_available_livreurs_api_v1_orders_livreurs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Get Available Livreurs Api V1 Orders Livreurs Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OrderCreate": {
        "properties": {
          "order_number": {
            "type": "string",
            "title": "Order Number"
          },
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "delivery_address": {
            "type": "string",
            "title": "Delivery Address"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus",
            "default": "En attente de validation"
          },
          "total_price": {
            "type": "integer",
            "title": "Total Price",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "livreur_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Livreur Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemBase"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "order_number",
          "client_name",
          "phone",
          "delivery_address",
          "items"
        ],
        "title": "OrderCreate"
      },
      "OrderItemBase": {
        "properties": {
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "integer",
            "title": "Unit Price"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "quantity",
          "unit_price"
        ],
        "title": "OrderItemBase"
      },
      "OrderRead": {
        "properties": {
          "order_number": {
            "type": "string",
            "title": "Order Number"
          },
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "delivery_address": {
            "type": "string",
            "title": "Delivery Address"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus",
            "default": "En attente de validation"
          },
          "total_price": {
            "type": "integer",
            "title": "Total Price",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "livreur_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Livreur Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemBase"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "order_number",
          "client_name",
          "phone",
          "delivery_address",
          "id",
          "items"
        ],
        "title": "OrderRead"
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "En attente de validation",
          "Valid\u00e9e - En pr\u00e9paration",
          "En transit",
          "Livr\u00e9",
          "Refus\u00e9e"
        ],
        "title": "OrderStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login/access-token"
          }
        }
      }
    }
  }
}